Multiple fixes:
1. Fix a critical file descriptor leak that would break when converting
   large/many files due to missing GC on os.open/tmpfile FDs
2. Make catchup more aggressive to help us when faced with unreliable
   servers.
3. Make conversion interruptible and keep original revision statistics.

+++ src/backy/backup.py
@@ -233,31 +233,39 @@ class Backup(object):
 
     @locked()
     def upgrade(self):
-        """Upgrade this backups store from cowfile to chunked.
+        """Upgrade this backup's store from cowfile to chunked.
 
         This can take a long time and is intended to be interruptable.
 
-        We start creating new backups with chunked right away to avoid
-        having to convert more in the future.
+        We start creating new backups with the new format once everything
+        is converted as we do not want to interfere with the config file
+        but allow upgrading without a format specification.
 
         """
-        # Phase 1: update config file to chunked
-        # XXX
-        # Phase 2: convert existing cowfiles
         from backy.backends.chunked import ChunkedFileBackend
         from backy.sources.file import File
-
         for revision in self.history:
             if revision.backend_type != 'cowfile':
                 continue
             print("Converting {}".format(revision.uuid))
             original_file = revision.filename + '.old'
-            os.rename(revision.filename, original_file)
+            if not os.path.exists(original_file):
+                # We may be resuming a partial upgrade. Only move the file if
+                # our .old doesn't exist.
+                os.rename(revision.filename, original_file)
+            else:
+                print("Resuming with existing .old")
             revision.writable()
             chunked = ChunkedFileBackend(revision)
             chunked.clone_parent = False
             cowfile = File(dict(filename=original_file, cow=False))(revision)
+            # Keep a copy of the statistics as it will get replaced when
+            # running the full copy.
+            original_stats = revision.stats.copy()
             cowfile.backup(chunked)
+            revision.stats = original_stats
+            revision.backend_type = 'chunked'
+            revision.write_info()
             revision.readonly()
             os.unlink(original_file)
